Restoration fail if the /var/www/$app don't exit
The app fails to restore if the /var/www/freshrss don't exits. Applied the latest helper to overcome it. I think it would solve the problem. But it needs to be tested. I will test it when I get the time.
Error log:
+ sudo chown -R root:root /var/www/freshrss
Warning: chown: cannot access ‘/var/www/freshrss’: No such file or directory
Error: Unable to restore the app 'freshrss'